Fix layout "jump" when switching between tabs
On the "Question List" tab, the content will "jump" a few pixels to the
left due to the content exceeding the page height (ie. it has a lot of
rows, which our table does). As a result, we need to account for this by
calculating the left and right margin when the user switches to the
respective tab, whether the scroll bar is present or not.

See
https://stackoverflow.com/questions/9341465/prevent-a-centered-layout-from-shifting-its-position-when-scrollbar-appears#answer-42987298